customView for Item_list

+++ JOA_Paris/app/src/main/java/com/sunah/foreignparis/korea/places/model/PlaceInfoModel.kt
@@ -5,8 +5,8 @@ class PlaceInfoModel
     val category: Int = 0
     var name: Map <String, String> = emptyMap()
     val address: String = ""
-    val latitude: Int = 0
-    val longitude: Int = 0
+    val latitude: Double = 0.0
+    val longitude: Double = 0.0
     val phone: String = ""
     val picture: String= ""
     val foursquare: String= ""
